refactor (test.user.repository: Optional 코드 수정 (#8)
- 기존 코드인 .getId()는 .isPresent()와 함께 사용해야 했음.
이미 생성된 기본 객체를 찾아 제공하는 경우이므로, Optional.orElse(null)를 통해 코드를 간결하게 변경.
